integrate cache manager with service pool
