+++ concert_service_manager/src/concert_service_manager/__init__.py
@@ -11,5 +11,4 @@ from .service_instance import *
 from .service_pool import ServicePool
 from .exceptions import InvalidSolutionConfigurationException, InvalidServiceProfileException
 from .service_profile import ServiceProfile
-from .service_cache_manager import ServiceCacheManager
 from .utils import *
