+++ concert_service_manager/src/concert_service_manager/service_profile.py
@@ -8,17 +8,19 @@
 import os
 import time
 import yaml
+import copy
 
-import concert_msgs.msg as concert_msgs
+
 import genpy
-import rocon_python_utils
 import rospkg
 import rospy
 import unique_id
-import rocon_console.console as console
-import scheduler_msgs.msg as scheduler_msgs
 
-from .exceptions import InvalidServiceProfileException
+import rocon_python_utils
+import scheduler_msgs.msg as scheduler_msgs
+import rocon_std_msgs.msg as rocon_std_msgs
+import concert_msgs.msg as concert_msgs
+from .utils import *
 
 ##############################################################################
 # Classes
@@ -27,157 +29,217 @@ from .exceptions import InvalidServiceProfileException
 
 class ServiceProfile(object):
     __slots__ = [
-            'msg',              # basic data storage for the profile - concert_msgs.ServiceProfile
-            '_last_modified',   # timestamp of last .service file modification
-            'hash',             # hashlib.sha224 used by the service pool to track this profile
-            'resource_name',    # local copy of the resource name
-            '_overrides',       # any overrides that need to be applied after any loading
-            '_location_cache',  # pointer to the service profile location cache maintained by the service pool { resource_name : filename }
-            # aliases
-            'name'
-        ]
+        '_concert_name',  # concert name for loading service profile from cache
+        '_profile_files',  # a list of pair about service profile files and lasted modification time (*.service, *.parameters, *.interactions)
+        '_is_read_from_default',  # flag whether service profile read from default or not
+        '_service_profile_file',  # file path containing service description; "*.service" file path
+        '_overrides',  # any overrides that need to be applied after any loading
+        'service_profile',   # data storage for the profile as type of dictionary
+        'name',  # service name
+        'msg',  # basic data storage for the profile - concert_msgs.ServiceProfile
+        'enabled',  # enabled status
+    ]
 
-    def __init__(self, hash_id, resource_name, overrides, service_profile_location_cache):
-        """
-          Loads the service profile given the data provided by the solution configuration
-          (i.e. resource name + overrides). We provide an arg for the filename here even
-          though we could look it up ourselves since the service manager has probably already
-          cached it.
+    def __init__(self, concert_name, is_read_from_default, service_profile_file, overrides=[]):
+        self._concert_name = concert_name
+        self._is_read_from_default = is_read_from_default
+        self.service_profile = None
+        self._profile_files = []
 
-          :param hash_id: hash of the solution configuration data element (for tracking purposes)
-          :type hash_id: hashlib.sha224
-
-          :param resource_name: pkg/name identifying a concert service
-          :type resource_name: str
-
-          :param overrides: dictionary of the possible overridable variables for a service, key set is bounded
-          :type overrides: dic
-
-          :param service_profile_location_cache: this class will use and update the cache if it's own filename changes
-          :type service_profile_location_cache: { resource_name : os pathname }
-
-          :returns: the solution configuration data for services in this concert
-          :rtype: concert_msgs.ServiceProfile
-
-          :raises: :exc:`concert_service_manager.InvalidServiceProfileException` if the service profile yaml is invalid
-        """
-        self.hash = hash_id
-        self._location_cache = service_profile_location_cache
+        self._service_profile_file = service_profile_file
         self._overrides = overrides
-        self.resource_name = resource_name
-        self._last_modified = None  # gets updated when we load the profile
+
+        loaded_profile = {}
         try:
-            self.msg = self._load_profile()
-            self._last_modified = time.ctime(os.path.getmtime(self._filename()))
-        except (rospkg.ResourceNotFound, IOError):
-            raise InvalidServiceProfileException("could not find service profile [%s]" % self.resource_name)
-        self._validate()  # can also raise InvalidServiceProfileException
-        # aliases
-        self.name = self.msg.name
+            if self._is_read_from_default:
+                loaded_profile = self._read_service_profile_from_default()
+            else:
+                loaded_profile = self._read_service_profiles_from_cache()
+        except (rospkg.ResourceNotFound, IOError) as e:
+            raise e
 
-    def __str__(self):
-        s = ''
-        s += console.green + self.hash.hexdigest() + console.reset + '\n'
-        s += console.cyan + "     resource_name" + console.reset + ": " + console.yellow + "%s\n" % self.resource_name + console.reset
-        s += console.cyan + "     name" + console.reset + ": " + console.yellow + "%s\n" % self.msg.name + console.reset
-        s += console.cyan + "     description" + console.reset + ": " + console.yellow + "%s\n" % self.msg.description + console.reset
-        s += console.cyan + "     author" + console.reset + ": " + console.yellow + "%s\n" % self.msg.author + console.reset
-        s += console.cyan + "     priority" + console.reset + ": " + console.yellow + "%s\n" % self.msg.priority + console.reset
-        if self.msg.icon.resource_name:
-            s += console.cyan + "     icon" + console.reset + ": " + console.yellow + "%s\n" % self.msg.icon.resource_name + console.reset
-        s += console.cyan + "     launcher_type" + console.reset + ": " + console.yellow + "%s\n" % self.msg.launcher_type + console.reset
-        if self.msg.launcher:
-            s += console.cyan + "     launcher" + console.reset + ": " + console.yellow + "%s\n" % self.msg.launcher + console.reset
-        if self.msg.interactions:
-            s += console.cyan + "     interactions" + console.reset + ": " + console.yellow + "%s\n" % self.msg.interactions + console.reset
-        if self.msg.parameters:
-            s += console.cyan + "     parameters" + console.reset + ": " + console.yellow + "%s\n" % self.msg.parameters + console.reset
-        return s
+        self.service_profile = copy.deepcopy(loaded_profile)
+        self.name = self.service_profile['name']
+        self.msg = self._service_profile_to_msg(self.service_profile)
+        if 'enabled' in self.service_profile.keys():
+            self.enabled = self.service_profile['enabled']
+        else:
+            self.enabled = None
 
-    def _validate(self):
+    def _check_modification(self):
         """
-        Checks that the service profile is a valid profile.
+        Check modification of service profile files.
 
-          :raises: :exc:`concert_service_manager.InvalidServiceProfileException` if the service profile yaml is invalid
+        :return: result of modification checking. If they are removed or modified, return true.
+        :rtype: bool 
         """
-        if self.msg.launcher_type != concert_msgs.ServiceProfile.TYPE_ROSLAUNCH and \
-           self.msg.launcher_type != concert_msgs.ServiceProfile.TYPE_SHADOW and \
-           self.msg.launcher_type != concert_msgs.ServiceProfile.TYPE_CUSTOM:
-            raise InvalidServiceProfileException("invalid service launcher type [%s][%s]" % (self.msg.launcher_type, self.resource_name))
+        is_change = False
+        for profile_file in self._profile_files:
+            file_name = profile_file[0]
+            modified_time = profile_file[1]
+            if not rocon_python_utils.ros.is_validation_file(file_name):
+                is_change = True
+            elif modified_time != time.ctime(os.path.getmtime(file_name)):
+                is_change = True
+                profile_file[1] = time.ctime(os.path.getmtime(file_name))
+        return is_change
 
-    def _load_profile(self):
+    def reload(self):
         """
-          :raises: :exc:`rospkg.ResourceNotFound` if the service profile yaml is not found
-          :raises: :exc:`concert_service_manager.InvalidServiceProfileException` if the service profile yaml is invalid
+        Reload the service profile from cached or default file as flag. 
+        However, enabled status of service is not update.
+
+        :raises: :exc:`rospkg.ResourceNotFound` if the service profile is not available
         """
-        msg = concert_msgs.ServiceProfile()
+        self._loginfo("detect [%s] service modification. reload" % self.name)
+        loaded_profile = {}
         try:
-            filename = self._filename()  # this can raise ResourceNotFound, IOError
-            with open(filename) as f:
-                loaded_profile = yaml.load(f)
+            if self._is_read_from_default:
+                loaded_profile = self._read_service_profile_from_default()
+            else:
+                loaded_profile = self._read_service_profiles_from_cache()
         except rospkg.ResourceNotFound as e:
             raise e
 
-        loaded_profile['resource_name'] = self.resource_name
+        self.service_profile = copy.deepcopy(loaded_profile)
+        self.name = self.service_profile['name']
+        self.msg = self._service_profile_to_msg(self.service_profile)
+
+    def _service_profile_to_msg(self, loaded_profile):
+        """
+        Change service proflies data to ros message
+
+        :returns: generated service profile message
+        :rtype: [concert_msgs.ServiceProfile]
+
+        """
+        msg = concert_msgs.ServiceProfile()
+        msg.uuid = unique_id.toMsg(unique_id.fromRandom())
+        # todo change more nice method
+        if 'resource_name' in loaded_profile:
+            msg.resource_name = loaded_profile['resource_name']
+        if 'name' in loaded_profile:
+            msg.name = loaded_profile['name']
+        if 'description' in loaded_profile:
+            msg.description = loaded_profile['description']
+        if 'author' in loaded_profile:
+            msg.author = loaded_profile['author']
+        if 'priority' in loaded_profile:
+            msg.priority = loaded_profile['priority']
+        if 'launcher_type' in loaded_profile:
+            msg.launcher_type = loaded_profile['launcher_type']
+        if 'icon' in loaded_profile:
+            msg.icon = rocon_python_utils.ros.icon_resource_to_msg(loaded_profile['icon'])
+        if 'launcher' in loaded_profile:
+            msg.launcher = loaded_profile['launcher']
+        if 'interactions' in loaded_profile:
+            msg.interactions = loaded_profile['interactions']
+        if 'parameters' in loaded_profile:
+            msg.parameters = loaded_profile['parameters']
+        if 'parameters_detail' in loaded_profile:
+            for param_key in loaded_profile['parameters_detail'].keys():
+                msg.parameters_detail.append(rocon_std_msgs.KeyValue(param_key, str(loaded_profile['parameters_detail'][param_key])))
+        return msg
+
+    def _read_service_profile_from_default(self):
+        """
+          Load service profile from resource file.
+
+          :returns: file path of loaded service profile
+          :rtype: str
+
+          :raises: :exc:`rospkg.ResourceNotFound` 
+
+        """
+        # load service profile from resource
+        loaded_profile = {}
+        service_file_name = self._service_profile_file
+        overrides = copy.deepcopy(self._overrides)
+
+        try:
+            file_name = rocon_python_utils.ros.find_resource_from_string(service_file_name)
+            with open(file_name) as f:
+                loaded_profile = yaml.load(f)
+            self._profile_files.append([file_name, time.ctime(os.path.getmtime(file_name))])
+        except rospkg.ResourceNotFound as e:
+            raise e
+        loaded_profile['resource_name'] = service_file_name
 
         # set priority to default if it was not configured
         if 'priority' not in loaded_profile.keys():
             loaded_profile['priority'] = scheduler_msgs.Request.DEFAULT_PRIORITY
-
-        # override parameters
         for key in loaded_profile:
-            if key in self._overrides and self._overrides[key] is not None:
-                loaded_profile[key] = self._overrides[key]
-
+            if key in overrides and overrides[key] is not None:
+                loaded_profile[key] = overrides[key]
         if 'launcher_type' not in loaded_profile.keys():  # not set
             loaded_profile['launcher_type'] = concert_msgs.ServiceProfile.TYPE_SHADOW
-        # We need to make sure the service description name is a valid rosgraph namespace name
-        # @todo Actually call the rosgraph function to validate this (do they have converters?)
-        loaded_profile['name'] = loaded_profile['name'].lower().replace(" ", "_")
+        loaded_profile['name'] = rocon_python_utils.ros.get_ros_friendly_name(loaded_profile['name'])
 
-        # load icon if necessary
-        replace = {}
-        if 'icon' in loaded_profile:
-            replace[loaded_profile['icon']] = rocon_python_utils.ros.icon_resource_to_msg(loaded_profile['icon'])
+        if 'parameters' in loaded_profile.keys():
+            loaded_profile['parameters_detail'] = []
+            try:
+                parameters_yaml_file = rocon_python_utils.ros.find_resource_from_string(rocon_python_utils.ros.check_extension_name(loaded_profile['parameters'], '.parameters'))
+                with open(parameters_yaml_file) as f:
+                    parameters_yaml = yaml.load(f)
+                    loaded_profile['parameters_detail'] = parameters_yaml
+                self._profile_files.append([parameters_yaml_file, time.ctime(os.path.getmtime(parameters_yaml_file))])
+            except rospkg.ResourceNotFound as e:
+                raise e
 
-        # generate message
-        genpy.message.fill_message_args(msg, loaded_profile, replace)
+        if 'interactions' in loaded_profile.keys():
+            try:
+                interactions_yaml_file = rocon_python_utils.ros.find_resource_from_string(rocon_python_utils.ros.check_extension_name(loaded_profile['interactions'], '.interactions'))
+                with open(interactions_yaml_file) as f:
+                    interactions_yaml = yaml.load(f)
+                    loaded_profile['interactions_detail'] = interactions_yaml
+                self._profile_files.append([interactions_yaml_file, time.ctime(os.path.getmtime(interactions_yaml_file))])
+            except rospkg.ResourceNotFound as e:
+                raise e
 
-        # fill in unique identifier used by services and their requesters
-        msg.uuid = unique_id.toMsg(unique_id.fromRandom())
+        return loaded_profile
 
-        return msg
-
-    def reload(self):
+    def _read_service_profiles_from_cache(self):
         """
-          Attempt to reload the profile from file if it has changed.
+        Load service profile from cached solution configuration.
 
-          :raises: :exc:`concert_service_manager.InvalidServiceProfileException` if the service profile yaml is invalid
+        :returns: file path of loaded service profile
+        :rtype: str
+
+        :raises: :exc:`rospkg.ResourceNotFound`
+
         """
-        modified_time = time.ctime(os.path.getmtime(self._filename()))
-        try:
-            if modified_time != self._last_modified:
-                self.msg = self._load_profile()
-                self._last_modified = time.ctime(os.path.getmtime(self._filename()))
-        except (rospkg.ResourceNotFound, IOError):
-            raise InvalidServiceProfileException("could not find service profile [%s]" % self.resource_name)
+        loaded_profile = {}
+        concert_name = self._concert_name
+        service_file_name = self._service_profile_file
 
-    def _filename(self):
-        '''
-          Try and find the service profile filename. Update the cache if necessary.
+        if not rocon_python_utils.ros.is_validation_file(service_file_name):
+            raise rospkg.ResourceNotFound("can not find service file in cache [%s]" % service_file_name)
+        else:
+            self._profile_files.append([service_file_name, time.ctime(os.path.getmtime(service_file_name))])
+            with open(service_file_name) as f:
+                loaded_profile = yaml.load(f)
+                if 'parameters' in loaded_profile.keys():
+                    loaded_profile['parameters_detail'] = []
+                    parameters_yaml_file = os.path.join(get_service_profile_cache_home(concert_name, loaded_profile['name']), loaded_profile['parameters'])
+                    if not rocon_python_utils.ros.is_validation_file(parameters_yaml_file):
+                        raise rospkg.ResourceNotFound("can not find parameters file in cache [%s]" % parameters_yaml_file)
+                    self._profile_files.append([parameters_yaml_file, time.ctime(os.path.getmtime(parameters_yaml_file))])
+                    with open(parameters_yaml_file) as f:
+                        parameters_yaml = yaml.load(f)
+                        loaded_profile['parameters_detail'] = parameters_yaml
+                if 'interactions' in loaded_profile.keys():
+                    loaded_profile['interactions_detail'] = []
+                    interactions_yaml_file = os.path.join(get_service_profile_cache_home(concert_name, loaded_profile['name']), loaded_profile['interactions'])
+                    if not rocon_python_utils.ros.is_validation_file(interactions_yaml_file):
+                        raise rospkg.ResourceNotFound("can not find interactions file in cache [%s]" % interactions_yaml_file)
+                    self._profile_files.append([interactions_yaml_file, time.ctime(os.path.getmtime(interactions_yaml_file))])
+                    with open(interactions_yaml_file) as f:
+                        interactions_yaml = yaml.load(f)
+                        loaded_profile['interactions_detail'] = interactions_yaml
+        return loaded_profile
 
-          :raises: :exc:`rospkg.ResourceNotFound` if the service profile yaml is not found
-        '''
-        # 1st attempt - look in da cache
-        try:
-            if os.path.isfile(self._location_cache[self.resource_name]):
-                return self._location_cache[self.resource_name]
-        except KeyError:
-            pass
-        # 2nd attempt - do the hard yakka
-        try:
-            filename = rocon_python_utils.ros.find_resource_from_string(self.resource_name)
-            self._location_cache[self.resource_name] = filename  # update the cache
-            return filename
-        except rospkg.ResourceNotFound:
-            raise rospkg.ResourceNotFound("could not find service profile [%s]" % self.resource_name)
+    def _loginfo(self, msg):
+        rospy.loginfo("Service Manager : " + str(msg))
+
+    def _logwarn(self, msg):
+        rospy.logwarn("Service Manager : " + str(msg))
